WarNiNg

+++ libgen/gen.h
@@ -23,8 +23,8 @@ namespace gen {
 	omnn::math::Valuable BuildFormula(const std::string& s, const omnn::math::Valuable::vars_cont_t& knownValues, const omnn::math::Valuable::va_names_t& varNames = InitialVarNames());
 	omnn::math::Valuable BuildFormulaAGRBofWidthHeightXYiN(uint32_t w, uint32_t h, omnn::math::Valuable r, omnn::math::Valuable g, omnn::math::Valuable b);
 
-	bool Generate(const omnn::math::Valuable& pattern, float* data, unsigned n);
-	bool Generate(const omnn::math::Valuable& pattern, uint32_t* data, unsigned n);
+	bool Generate(const omnn::math::Valuable& pattern, float* data, size_t n);
+	bool Generate(const omnn::math::Valuable& pattern, uint32_t* data, size_t n);
 	
 	boost::gil::rgba8_image_t GeneratePatternImage(uint32_t width, uint32_t height, omnn::math::Valuable red, omnn::math::Valuable green, omnn::math::Valuable blue);
 }
